+++ src/search.cpp
@@ -200,7 +200,7 @@ namespace stoat {
 
     void Searcher::startSearch(
         const Position& pos,
-        std::span<const u64> keyHistory,
+        std::span<const u128> keyHistory,
         util::Instant startTime,
         bool infinite,
         i32 maxDepth,
